+++ hosts/echo/darwin-configuration.nix
@@ -1,4 +1,4 @@
-{ keys, ... }:
+{ ... }:
 
 {
   networking.knownNetworkServices = [ "Ethernet" ];
@@ -14,8 +14,6 @@
     publicHostKey =
       "c3NoLWVkMjU1MTkgQUFBQUMzTnphQzFsWkRJMU5URTVBQUFBSUpCV2N4Yi9CbGFxdDFhdU90RStGOFFVV3JVb3RpQzVxQkorVXVFV2RWQ2Igcm9vdEBuaXhvcwo=";
   }];
-
-  nix.settings.trusted-public-keys = [ keys.signing.chi-linux-builder ];
 
   nix.settings.min-free = 3 * 1024 * 1024 * 1024;
 }