+++ hosts/sigma/hardware-configuration.nix
@@ -1,4 +1,4 @@
-{ config, lib, utils, ... }:
+{ config, lib, ... }:
 
 {
   boot.initrd.availableKernelModules =
@@ -58,33 +58,6 @@
         };
       };
     };
-  };
-
-  fileSystems."/mnt/phi" = {
-    device = "enzime@phi:/";
-    fsType = "fuse.sshfs";
-    noCheck = true;
-    options = [
-      "noauto"
-      "x-systemd.automount"
-      "_netdev"
-      "IdentityFile=/etc/ssh/ssh_host_ed25519_key"
-      "allow_other"
-      "uid=1000"
-      "gid=100"
-      "ConnectTimeout=1"
-      "x-systemd.mount-timeout=10s"
-      "ServerAliveInterval=1"
-      "ServerAliveCountMax=5"
-    ];
-  };
-
-  systemd.units."${utils.escapeSystemdPath "/mnt/phi"}.mount" = {
-    text = ''
-      [Unit]
-      StartLimitIntervalSec=0
-    '';
-    overrideStrategy = "asDropin";
   };
 
   powerManagement.cpuFreqGovernor = lib.mkDefault "powersave";