+++ modules/personal.nix
@@ -6,7 +6,34 @@
       builtins.attrValues { inherit (pkgs) apparency; };
   };
 
-  nixosModule = { user, pkgs, ... }: {
+  nixosModule = { user, pkgs, utils, ... }: {
+    fileSystems."/mnt/phi" = {
+      device = "enzime@phi:/";
+      fsType = "fuse.sshfs";
+      noCheck = true;
+      options = [
+        "noauto"
+        "x-systemd.automount"
+        "_netdev"
+        "IdentityFile=/etc/ssh/ssh_host_ed25519_key"
+        "allow_other"
+        "uid=1000"
+        "gid=100"
+        "ConnectTimeout=1"
+        "x-systemd.mount-timeout=10s"
+        "ServerAliveInterval=1"
+        "ServerAliveCountMax=5"
+      ];
+    };
+
+    systemd.units."${utils.escapeSystemdPath "/mnt/phi"}.mount" = {
+      text = ''
+        [Unit]
+        StartLimitIntervalSec=0
+      '';
+      overrideStrategy = "asDropin";
+    };
+
     services.resilio.enable = true;
     services.resilio.listeningPort = 44444;
     services.resilio.enableWebUI = true;